docs: 設計ドキュメント（Mermaid 図つき）を追加し、対応 DB 表を実測で更新
docs/architecture.md を新規追加。「どこに何があり、なぜそうなっているか」を
初見の開発者向けに Mermaid 図 11 枚で説明する:

- 全体像（1 プロセス構成）とパッケージ依存の向き（禁止エッジを赤破線 + ラベルで明示）
- アダプター層の契約（classDiagram）と値のワイヤー形式
- セッションと接続プールの状態遷移
- 行の閲覧 / SQL 実行（ストリーミング・キャンセル）/ DDL プレビューのシーケンス
- エクスポートのセクション順序とインポートパイプライン
- 品質ゲート（ローカル hooks → CI）と「どこを変更するか」の逆引き表

各図は mermaid-cli で実際にレンダリングして検証済み。

あわせて:
- 対応データベース表を実測値に更新。MySQL 9・Percona Server 8.4・PostgreSQL 18 を
  統合テスト全通過として追加し、互換エンジン（TiDB 7.5 = 113 件中 82 件、
  CockroachDB 24.3 = 55 件）の実測結果と判定を明記
- README の「動作要件」にあった古い検証済みバージョン記述を deployment.md への
  参照に置き換え（バージョン情報の唯一の正を 1 か所に）
- PostgreSQL の行数推定で reltuples / n_live_tup を ::float8 にキャスト
  （CockroachDB では decimal を返すため型不一致になっていた。実 PG では挙動不変）

+++ packages/adapter/src/postgres/introspect.ts
@@ -50,7 +50,7 @@ export async function pgListTables(conn: Conn, ns: Namespace): Promise<TableInfo
   const r = firstResult(
     await conn.query(
       `SELECT c.relname, c.relkind,
-              CASE WHEN c.reltuples < 0 THEN s.n_live_tup ELSE c.reltuples END,
+              CASE WHEN c.reltuples < 0 THEN s.n_live_tup::float8 ELSE c.reltuples::float8 END,
               obj_description(c.oid, 'pg_class'),
               CASE WHEN c.relkind IN ('r', 'p', 'm') THEN pg_total_relation_size(c.oid) END,
               (SELECT string_agg(p.relname, $2 ORDER BY i.inhseqno) FROM pg_inherits i JOIN pg_class p ON p.oid = i.inhparent
